align code with changes from unique-ids commit 5af449d from itsmojo (#5)
* align code with changes from unique-ids commit 5af449d from itsmojo
Unique controller ids support, working eap seq # podIds
* persistent id created at startup or deactvation as needed
* eap seq # now incremented for each new session
* podId #'s now cycle between next of 3 values on deactivate
* temp fake podState no longer used to implement pairing

+++ OmniBLE/PumpManagerUI/Views/PodDetailsView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 import LoopKitUI
 
 public struct PodVersion {
-    var lotNumber: UInt64
+    var lotNumber: UInt32
     var sequenceNumber: UInt32
     var firmwareVersion: String
     var bleFirmwareVersion: String
